complete tests in sql/rabbitmq

+++ api/models/permission_user_model.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import BigInteger, Column, ForeignKey
+from sqlalchemy import BigInteger, Column, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql.sqltypes import Boolean
 
@@ -10,7 +10,7 @@ from common.base_model import BaseModelInterface
 class PermissionUserModel(BaseModelInterface):
     __tablename__ = 'permission_user_model'
 
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
 
     user_id = Column(
         BigInteger, ForeignKey('user_model.id', ondelete='CASCADE')
